+++ src/main/java/com/example/demo/utils/DataLoaderInterface.java
@@ -6,6 +6,5 @@ import java.util.Map;
 import org.json.simple.parser.ParseException;
 
 public interface DataLoaderInterface {
-
-    Map<String, Integer> loadData() throws IOException, ParseException;
+    Map<String, Long> loadData() throws IOException, ParseException;
 }
